+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Head
 
+- Added: `font-family-name-quotes` rule.
 - Fixed: bug in `rule-properties-order` empty line detection when the two newlines were separated
   by some other whitespace.
 - Fixed: option parsing bug that caused problems when using the `"alphabetical"` primary option